refactor(web-ui): decompose OrderReviewExperience god-component
Split the 875-line OrderReviewExperience into three files:
- orderReviewHelpers.tsx: MetricTile, EmptySection, classifyInvalidationRule,
  countryFromTicker, REVIEW_SECTIONS, country-suffix map, ReviewSectionId
- OrderReviewSummary.tsx: the decision/setup/risk review carousel, fully
  presentational, takes derived data via props
- OrderReviewExperience.tsx: container (form, react-hook-form state, risk
  metrics, submit handler), now 568 lines, renders OrderReviewSummary

OrderReviewContext stays exported from the container. No behavior or public
interface change.
